商品購入_r5_2：登録したカードから購入、商品購入確認画面フロント、マイページからカード登録
