+++ app/assets/stylesheets/modules/_purchase.scss
@@ -87,6 +87,7 @@
         }
       }
       &__credit{
+        text-align: left;
         &__link{
           text-decoration: none;
           color: deepskyblue;
