+++ app/controllers/items_controller.rb
@@ -1,6 +1,6 @@
 class ItemsController < ApplicationController
-  before_action :set_item, only: [:show, :purchase, :pay]
-  before_action :set_card, only: [:purchase, :pay]
+  before_action :set_item, only: [:show, :purchase, :pay, :card_show]
+  before_action :set_card, only: [:purchase, :pay, :card_show]
 
   def index
     @items = Item.includes(:item_imgs).order('created_at DESC')
@@ -30,7 +30,6 @@ class ItemsController < ApplicationController
       customer = Payjp::Customer.retrieve(@card.customer_id)
       @default_card_information = customer.cards.retrieve(@card.card_id)
     end
-    # render template:'cards/new'(確認中)
   end
 
   def pay
@@ -47,10 +46,17 @@ class ItemsController < ApplicationController
   end
 
   def card_show
+    if @card.blank?
+      redirect_to action: "new" 
+    else
+      Payjp.api_key = ENV["PAYJP_PRIVATE_KEY"]
+      customer = Payjp::Customer.retrieve(@card.customer_id)
+      @default_card_information = customer.cards.retrieve(@card.card_id)
+    end
   end
 
   def set_item
-    @item = Item.find_by(params[:id])
+    @item = Item.find_by(id:params[:id])
   end
 
   def set_card
