+++ app/views/items/purchase.html.haml
@@ -25,17 +25,10 @@
       %ul.purchase__body__way-to-pay__title
         %li
           支払い方法
-        -# = link_to delete_cards_path ,method: :post,class:"purchase__body__sending-destination__title__link" do
-        -#   -# %button{type: "submit",class: "btn-red"}
-          削除する 〉
-        -# = form_tag(delete_cards_path, method: :post, id: 'charge-form',  name: "inputForm") do
-        -#   %input{ type: "hidden", name: "card_id", value: "" }
-        -#   %button 削除する
       .purchase__body__way-to-pay__credit
         - if @card
-          = render template: 'cards/show'
+          = render template: 'items/card_show' 
         - else
-          -# = render template: 'cards/new'
           = link_to new_card_path(@item) ,method: :get,class:"purchase__body__way-to-pay__credit__link" do
             ※登録してください
     .purchase__body__sending-destination
@@ -44,7 +37,7 @@
           配送先
         = link_to "http://localhost/#", class: "purchase__body__sending-destination__title__link" do
           変更する 〉
-          -# (以下要修正)
+        -# (以下確認中)
         -# = form_tag(new_send_path, method: :post, id: 'charge-form',  name: "inputForm") do
         -#   %input{ type: "hidden", name: "card_id", value: "" }
         -#   %button 変更する
