--- app/views/purchases/done.html.haml
@@ -1,3 +1,0 @@
-%h2 購入が完了しました！
-%p Apple MacBook Pro 13インチ
-%p ¥135,000(送料込み)
