--- app/views/users/regist.html.haml
@@ -1,188 +1,0 @@
-%body
-  %header.pcHeader
-    .headerInner
-      .mainHeader
-        %h1.icon
-          = link_to root_path do
-            = image_tag 'logo.png', height: '41', width: '140'
-        .searchBox
-          %form{"accept-charset" => "UTF-8", action: "#", method: "get"}
-            %input{name: "utf8", type: "hidden", value: "✓"}/
-            %input#searchInput.searchBox--input{name: "key", placeholder: "キーワードから探す"}/
-            %button{type: "submit"}
-              .searchBox__icon
-                = image_tag 'icon-search 1.png', height: '20', width: '20'
-            %input{name: "authenticity_token", type: "hidden", value: "kRlCZacZocQxRbMm04ok7Aze5BEkOhq1nVZ4OizLi+k7nVGGHdamQKYlgribxg0EitLTAzfJGJ3VOvckwE+uEA=="}/
-        %ul.resultLists
-      .nav
-        %ul.listsLeft
-          %li.listsLeft__item.listsLeft__item--first
-            = link_to "#", class:"#catBtn" do
-              カテゴリー
-            .tree.displayNone
-              %ul.categoryTree
-                %li.categoryTree--item
-                  = link_to "#", class:"#1" do
-                    レディース
-                %li.categoryTree--item
-                  = link_to "#", class:"#2" do
-                    メンズ
-                %li.categoryTree--item
-                  = link_to "#", class:"#3" do
-                    ベビー・キッズ
-                %li.categoryTree--item
-                  = link_to "#", class:"#4" do
-                    インテリア・住まい・小物
-                %li.categoryTree--item
-                  = link_to "#", class:"#5" do
-                    本・音楽・ゲーム
-                %li.categoryTree--item
-                  = link_to "#", class:"#6" do
-                    おもちゃ・ホビー・グッズ
-                %li.categoryTree--item
-                  = link_to "#", class:"#7" do
-                    コスメ・香水・美容
-                %li.categoryTree--item
-                  = link_to "#", class:"#8" do
-                    家電・スマホ・カメラ
-                %li.categoryTree--item
-                  = link_to "#", class:"#9" do
-                    スポーツ・レジャー
-                %li.categoryTree--item
-                  = link_to "#", class:"#10" do
-                    ハンドメイド
-                %li.categoryTree--item
-                  = link_to "#", class:"#11" do
-                    チケット
-                %li.categoryTree--item
-                  = link_to "#", class:"#12" do
-                    自動車・オートバイ
-                %li.categoryTree--item
-                  = link_to "#", class:"#13" do
-                    その他
-          %li.listsLeft__item
-            = link_to "#", class:"#brandBtn" do
-              ブランド
-            %ul.brandsPulldown.displayNone
-        %ul.listsRight
-          - if user_signed_in?
-            %li.listsRight__item.listsRight__item--login
-              = link_to "ログアウト", destroy_user_session_path, method: :delete
-          - else
-            %li.listsRight__item.listsRight__item--login
-              = link_to "ログイン", new_user_session_path
-            %li.listsRight__item.listsRight__item--new
-              = link_to "新規会員登録", new_user_registration_path
-
-%main.l-container.clearfix
-  .l-side
-    %nav.mypage-nav
-      %ul.mypage-nav-list
-        %li
-          -# (以下パスのプレフィックス指定ではNameError)
-          = link_to "/jp/mypage/", class:"mypage-nav-list-item.active" do
-            マイページ
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/notification/", class:"mypage-nav-list-item" do
-            お知らせ
-            %span.mypage-nav-number
-              99+
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/todo/", class:"mypage-nav-list-item" do
-            やることリスト
-            %span.mypage-nav-number
-              1
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/like/history/", class:"mypage-nav-list-item" do
-            いいね！一覧
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/sell/", class:"mypage-nav-list-item" do
-            出品する
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/drafts/", class:"mypage-nav-list-item" do
-            下書き一覧
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/listings/listing/", class:"mypage-nav-list-item" do
-            出品した商品 - 出品中
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/listings/in_progress/", class:"mypage-nav-list-item" do
-            出品した商品 - 取引中
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/listings/completed/", class:"mypage-nav-list-item" do
-            出品した商品 - 売却済み
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/purchase/", class:"mypage-nav-list-item" do
-            購入した商品 - 取引中
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/purchased/", class:"mypage-nav-list-item" do
-            購入した商品 - 過去の取引
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/news/", class:"mypage-nav-list-item" do
-            ニュース一覧
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/review/history/", class:"mypage-nav-list-item" do
-            評価一覧
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/help_center/", class:"mypage-nav-list-item" do
-            ガイド
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/support/", class:"mypage-nav-list-item" do
-            お問い合わせ
-            %i.icon-arrow-right
-      %h3.mypage-nav-head メルペイ
-      %ul.mypage-nav-list
-        %li
-          = link_to "/jp/mypage/sales/", class:"mypage-nav-list-item" do
-            残高・振込申請
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/point/", class:"mypage-nav-list-item" do
-            ポイント
-            %i.icon-arrow-right
-      %h3.mypage-nav-head 設定
-      %ul.mypage-nav-list
-        %li
-          = link_to "/jp/mypage/profile/", class:"mypage-nav-list-item" do
-            プロフィール
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/deliver_address/", class:"mypage-nav-list-item" do
-            発送元・お届け先住所変更
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/card/", class:"mypage-nav-list-item" do
-            支払い方法
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/email_password/", class:"mypage-nav-list-item" do
-            メール/パスワード
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/identification/", class:"mypage-nav-list-item" do
-            本人情報
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/mypage/sms_confirmation/", class:"mypage-nav-list-item" do
-            電話番号の確認
-            %i.icon-arrow-right
-        %li
-          = link_to "/jp/logout/", class:"mypage-nav-list-item" do
-            ログアウト
-            %i.icon-arrow-right
-  .l-content
-    .__credit
-      = render template: 'cards/new'
